Improvements to VideoClipsGluer: glues together front videos into on front video, left into one left, and right into one right

+++ VideoClipsGluer/Program.cs
@@ -15,7 +15,7 @@ namespace VideoClipsGluer
             string workingDir;
             if (args.Length < 1 || args[0] == "")
             {
-                workingDir = "/Users/dstrube/Downloads/TeslaCam/SavedClips/";
+                workingDir = "/Users/dstrube/Downloads/TeslaCam/SavedClips";
             }
             else
             {
@@ -25,64 +25,33 @@ namespace VideoClipsGluer
             var dirs = new ArrayList(Directory.GetDirectories(workingDir));
             dirs.Sort();
             dirs.TrimToSize();
-            ProcessWindowStyle windowStyle = ProcessWindowStyle.Hidden;//Normal;              
-            Process p1a = new Process();
-            try
+            foreach (var dir in dirs)
             {
-                foreach (var dir in dirs)
+                Console.WriteLine($"dir: {dir}");
+                var mp4files = new ArrayList(Directory.GetFiles(dir.ToString(), "*.mp4"));
+                //processFileType("all", mp4files, dir.ToString());
+
+                var mp4filesFront = new ArrayList(Directory.GetFiles(dir.ToString(), "*-front.mp4"));
+                processFileType("front", mp4filesFront, dir.ToString());
+
+                var mp4filesLeft = new ArrayList(Directory.GetFiles(dir.ToString(), "*-left_repeater.mp4"));
+                processFileType("left", mp4filesLeft, dir.ToString());
+
+                var mp4filesRight = new ArrayList(Directory.GetFiles(dir.ToString(), "*-right_repeater.mp4"));
+                processFileType("right", mp4filesRight, dir.ToString());
+
+                if (mp4files.Count == 1)
                 {
-                    Console.WriteLine($"dir: {dir}");
-                    var mp4files = new ArrayList(Directory.GetFiles(dir.ToString(), "*.mp4"));
-                    mp4files.TrimToSize();
-                    mp4files.Sort();
-                    if (mp4files.Count == 1)
-                    {
-                        //there is only 1 so skip this folder
-                        continue;
-                    }
-                    //TODO Only combine files from the same day
-                    //If multiple days exist in 1 folder, make a separate file for each day
-                    Console.WriteLine($"file count: {mp4files.Count}");
-                    Console.WriteLine("List of files:");
-                    var stringBuilder = new StringBuilder();
-                    foreach (var file in mp4files)
-                    {
-                        FileInfo fileInfo = new FileInfo(file.ToString());
-                        if (fileInfo.Length == 0)
-                        {
-                            Console.WriteLine($"Empty file; deleting: {file}");
-                            File.Delete(file.ToString());
-                            continue;
-                        }
-                        var content = $"file {file.ToString().Substring(1 + file.ToString().LastIndexOf(Path.DirectorySeparatorChar))}";
-                        Console.WriteLine(content);
-                        stringBuilder.AppendLine(content);
-
-                    }
-
-                    File.WriteAllText(dir.ToString() + Path.DirectorySeparatorChar + "input.txt", stringBuilder.ToString());
-
-                    Console.WriteLine($"ffmpeg -f concat -i {dir}/input.txt -c copy {dir}/output.mp4");
-
-                    p1a.StartInfo.FileName = "ffmpeg";
-                    p1a.StartInfo.WindowStyle = windowStyle;
-                    p1a.StartInfo.Arguments = $"-f concat -i {dir}{Path.DirectorySeparatorChar}input.txt -c copy {dir}{Path.DirectorySeparatorChar}output.mp4";
-                    p1a.Start();
-
-                    if (stopAtFirstDir)
-                    {
-                        break;
-                    }
+                    //there is only 1 so skip this folder
+                    continue;
+                }
+                if (stopAtFirstDir)
+                {
+                    break;
                 }
             }
-            finally
-            {
-                p1a.Dispose();
-            }
 
-            Console.WriteLine($"Hit enter to continue...");
-            Console.ReadLine();
-
+            /*
             Console.WriteLine($"Done gluing. Now deleting originals...");
             foreach (var dir in dirs)
             {
@@ -100,14 +69,61 @@ namespace VideoClipsGluer
                     {
                         continue;
                     }
-                    File.Delete(file.ToString());
+                    //File.Delete(file.ToString());
                 }
                 if (stopAtFirstDir)
                 {
                     break;
                 }
-            }
+            }*/
             Console.WriteLine($"Done");
+        }
+
+        static void processFileType(string fileType, ArrayList files, string dir)
+        {
+            files.TrimToSize();
+            files.Sort();
+            Console.WriteLine($"{fileType} file count: {files.Count}");
+            //TODO Only combine files from the same day
+            //If multiple days exist in 1 folder, make a separate file for each day
+            //Console.WriteLine("List of files:");
+            var stringBuilder = new StringBuilder();
+            foreach (var file in files)
+            {
+                FileInfo fileInfo = new FileInfo(file.ToString());
+                if (fileInfo.Length == 0)
+                {
+                    Console.WriteLine($"Empty file; deleting: {file}");
+                    File.Delete(file.ToString());
+                    continue;
+                }
+                var content = $"file {file.ToString().Substring(1 + file.ToString().LastIndexOf(Path.DirectorySeparatorChar))}";
+                //Console.WriteLine(content);
+                stringBuilder.AppendLine(content);
+
+            }
+
+            File.WriteAllText(dir + Path.DirectorySeparatorChar + fileType + "_input.txt", stringBuilder.ToString());
+
+            ProcessWindowStyle windowStyle = ProcessWindowStyle.Hidden;//Normal;              
+            Process p1a = new Process();
+            try
+            {
+                Console.WriteLine($"ffmpeg -f concat -i {dir + Path.DirectorySeparatorChar + fileType}_input.txt -c copy {dir + Path.DirectorySeparatorChar + fileType}_output.mp4");
+                p1a.StartInfo.FileName = "ffmpeg";
+                p1a.StartInfo.WindowStyle = windowStyle;
+                p1a.StartInfo.Arguments = $"-f concat -i {dir + Path.DirectorySeparatorChar + fileType}_input.txt -c copy {dir + Path.DirectorySeparatorChar + fileType}_output.mp4";
+                p1a.Start();
+                Console.WriteLine($"Hit enter after the file stuff is done...");
+                Console.ReadLine();
+                /*
+                */
+            }
+            finally
+            {
+                p1a.Dispose();
+            }
+
         }
     }
 }
